Fix some tests

+++ ss_stereo_spot5/validate.sh
@@ -22,8 +22,14 @@ cmp_stats.sh $file $gold
 gdalinfo -stats $file | grep -v Files | grep -v -i tif > run.txt
 gdalinfo -stats $gold | grep -v Files | grep -v -i tif > gold.txt
 
-diff=$(diff run.txt gold.txt)
-cat run.txt
+diff run.txt gold.txt
+
+max_err.pl run.txt gold.txt # print the error
+ans=$(max_err.pl run.txt gold.txt 1e-10) # compare the error
+if [ "$ans" -eq 0 ]; then
+    echo Validation failed
+    exit 1
+fi
 
 rm -f run.txt gold.txt
